Update README.md
The essence of this project revolves around creating an efficient data sorting system using a restricted set of commands and minimizing the number of operations required. To achieve this objective, the project will entail the careful orchestration of diverse sorting algorithms, strategically selecting the most suitable solutions to optimize the data sorting process.